[hail] Rename `MatrixTable.scala` and move to `is.hail.expr.ir`
