+++ hail/src/main/scala/is/hail/expr/ir/functions/RelationalFunctions.scala
@@ -1,12 +1,11 @@
 package is.hail.expr.ir.functions
 
-import is.hail.expr.ir.{ExecuteContext, LowerMatrixIR, MatrixValue, TableValue}
+import is.hail.expr.ir.{ExecuteContext, LowerMatrixIR, MatrixValue, RelationalSpec, TableValue}
 import is.hail.expr.types.virtual.Type
 import is.hail.expr.types.{BlockMatrixType, MatrixType, TableType}
 import is.hail.linalg.BlockMatrix
 import is.hail.methods._
 import is.hail.rvd.RVDType
-import is.hail.variant.RelationalSpec
 import org.json4s.ShortTypeHints
 import org.json4s.jackson.Serialization
